+++ src/main/java/com/telepathicgrunt/blame/main/FlatChunkGeneratorConfigBlame.java
@@ -20,13 +20,13 @@ public class FlatChunkGeneratorConfigBlame {
 
 	/**
 	 * Will detect if someone added the structure spacing to the chunkgenerator without adding
-	 * the structure to the FlatGenerationSettings's STRUCTURES map. Doing so will cause a crash.
+	 * the structure to the FlatChunkGeneratorConfig's STRUCTURE_FEATURE map. Doing so will cause a crash.
 	 * This mixin will say what structure it was that triggered it.
 	 */
-	public static void addCrashDetails(Map<StructureFeature<?>, ConfiguredStructureFeature<?, ?>> main_structure_map, Map.Entry<StructureFeature<?>, ConfiguredStructureFeature<?, ?>> structureEntry)
+	public static void addCrashDetails(Map<StructureFeature<?>, ConfiguredStructureFeature<?, ?>> mainStructureMap, Map.Entry<StructureFeature<?>, ConfiguredStructureFeature<?, ?>> structureEntry)
 	{
 		// This condition will cause a crash
-		if(main_structure_map.get(structureEntry.getKey()) == null){
+		if(mainStructureMap.get(structureEntry.getKey()) == null){
 
 			Identifier rl = Registry.STRUCTURE_FEATURE.getId(structureEntry.getKey());
 			String extraDetail = rl != null ? (" | " + rl.toString()) : "";
@@ -35,7 +35,7 @@ public class FlatChunkGeneratorConfigBlame {
 			String errorReport = "\n****************** Blame Report " + Blame.VERSION + " ******************" +
 					"\n\n A crash is most likely going to happen right after this report!" +
 					"\n It seems " + structureEntry.getKey().getName() + extraDetail + " is the cause because it is not added " +
-					"\n to the FlatGenerationSettings.STRUCTURES map. Please let the mod owner " +
+					"\n to the FlatChunkGeneratorConfig.STRUCTURE_FEATURE map. Please let the mod owner " +
 					"\n of that structure know about this crash. That way they can add their structure " +
 					"\n to that map since someone is trying to spawn it in a flat/custom dimension. \n\n";